feat: Update Angular routing module and navigation bar
The Angular routing module has been added to the project and the navigation bar has been updated to include the new routes. This commit also includes the generation of the routing module file with an array of routes, including the home, binding, directivas, pipes-angular, pipes-personalizados, and not-found components. The NavBar component's HTML template is also updated to include the new routes using the routerLink directive.